Fix logging implicit sign conversions
The warnings were being hidden by the use of -isystem to include
frameworks/native/include.

Bug: 31752268
Test: m -j
Merged-in: I2ce11db524ee9d8f846fdf3ed6fb882d5ef57956

+++ libs/gui/BufferQueueProducer.cpp
@@ -899,7 +899,8 @@ status_t BufferQueueProducer::queueBuffer(int slot,
                 mCore->mTransformHint,
                 static_cast<uint32_t>(mCore->mQueue.size()));
 
-        ATRACE_INT(mCore->mConsumerName.string(), mCore->mQueue.size());
+        ATRACE_INT(mCore->mConsumerName.string(),
+                static_cast<int32_t>(mCore->mQueue.size()));
 
         // Take a ticket for the callback functions
         callbackTicket = mNextCallbackTicket++;
